# src/main/java/com/openbanking/account/access/consents/model/OBReadConsent1.java
package com.openbanking.account.access.consents.model;

import java.io.Serializable;
import lombok.Getter;
import lombok.Setter;

/**
 * OBReadConsent1
 */
@Getter
@Setter
public class OBReadConsent1 implements Serializable {
	private static final long serialVersionUID = 1L;

}