# src/main/java/com/openbanking/account/access/consents/model/OBReadConsentResponse1.java
package com.openbanking.account.access.consents.model;

import java.io.Serializable;
import lombok.Getter;
import lombok.Setter;

/**
 * OBReadConsentResponse1
 */
@Getter
@Setter
public class OBReadConsentResponse1 implements Serializable {
	private static final long serialVersionUID = 1L;

}